cleanup some code.  now compiles on the simulator. prepping for custom sdks

+++ Source/HTTPRiot/HTTPRiotRequest.h
@@ -5,11 +5,11 @@
 //  Created by Justin Palmer on 1/30/09.
 //  Copyright 2009 Alternateidea. All rights reserved.
 //
-
+#import <Foundation/Foundation.h>
 #import "HTTPRiotConstants.h"
 
 @interface HTTPRiotRequest : NSObject {
-    CGFloat timeout;
+    float timeout;
     kHTTPRiotMethod httpMethod;
     NSString *path;
     NSDictionary *options;
@@ -17,7 +17,7 @@
     id formatter;
 }
 
-@property (nonatomic, assign) CGFloat timeout;
+@property (nonatomic, assign) float timeout;
 @property (nonatomic, assign) kHTTPRiotMethod httpMethod;
 @property (nonatomic, readonly, copy) NSString *path;
 @property (nonatomic, readonly, retain) NSDictionary *options;
